Update.
1999-12-03  Ulrich Drepper  <drepper@cygnus.com>

	* sysdeps/generic/bits/stropts.h: Update with LiS types and
	values.  Patch by David Grothe <dave@gcom.com>.

	* sysdeps/unix/sysv/linux/bits/types.h (__t_scalar_t, __t_uscalar_t):
	Make long types.
	* sysdeps/unix/sysv/linux/alpha/bits/types.h: Likewise.
	* sysdeps/unix/sysv/linux/mips/bits/types.h: Likewise.
	* sysdeps/unix/sysv/linux/sparc/bits/types.h: Likewise.

+++ sysdeps/unix/sysv/linux/alpha/bits/types.h
@@ -117,8 +117,8 @@ typedef struct
   } __fd_set;
 
 /* Used in XTI.  */
-typedef int __t_scalar_t;
-typedef unsigned int __t_uscalar_t;
+typedef long int __t_scalar_t;
+typedef unsigned long int __t_uscalar_t;
 
 /* Duplicates info from stdint.h but this is used in unistd.h.  */
 typedef long int __intptr_t;
